backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
openai>=1.0.0
pypdf>=5.0.0
python-docx>=1.0.0
